feat: Add uploadEventBanner method to FileRepositoryS3
This commit adds the uploadEventBanner method to the FileRepositoryS3 class. This method allows for uploading event banners to the S3 bucket. It takes in parameters such as the event ID, event name, image name key, event photo, and mimetype. The method uses the AWS SDK to upload the event banner to the specified S3 bucket.

Ref: #123